change order of test

+++ tests/end_to_end/eabi64/ctx.c
@@ -1,2 +1,2 @@
 void no_args_func();
-void func_with_args(float, float, int, int);
+void func_with_args(float, int, float, int);
